+++ client/src/app/admin/cadastros_auxiliares/categoria-analise.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: 'categoria-analise.component.html',
+})
+export class CategoriaAnaliseComponent implements OnInit {
+  constructor() {  }
+
+  ngOnInit() {}
+}