+++ client/src/app/admin/cadastros_auxiliares/usuario.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  templateUrl: 'usuario.component.html',
+})
+export class UsuarioComponent implements OnInit {
+  constructor() {  }
+
+  ngOnInit() {}
+}